refactor config and client
